LayoutTests/platform/ios-sim/iphone/fast/events/touch/document-create-touch-list-ios-expected.txt

This tests support for the document.createTouchList and window.TouchEvent APIs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'TouchEvent' in window is true
PASS window.propertyIsEnumerable('TouchEvent') is false
PASS new TouchEvent() threw exception TypeError: '[object TouchEventConstructor]' is not a constructor (evaluating 'new TouchEvent()').
PASS "createTouchList" in document is true
PASS touchList is non-null.
PASS touchList.length is 0
PASS touchList.item(0) is null
PASS touchList.item(1) is null
PASS successfullyParsed is true

TEST COMPLETE
